# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: CairnOps API
  version: 0.1.0-connected
  description: API CairnOps pour la supervision native et les Connecteurs guidés.
servers:
  - url: /
components:
  securitySchemes:
    BootstrapBearer:
      type: http
      scheme: bearer
      description: >
        Jeton défini par CAIRNOPS_BOOTSTRAP_TOKEN. Il n'ouvre que deux portes, toutes
        deux hors session : l'initialisation de l'instance et le rétablissement d'un
        accès perdu. Il prouve le contrôle du déploiement, rien d'autre.
    SessionCookie:
      type: apiKey
      in: cookie
      name: cairnops_session
      description: Session HttpOnly révocable. Sous HTTPS, le serveur utilise le préfixe de sécurité __Host-.
    DeviceBearer:
      type: http
      scheme: bearer
      bearerFormat: CairnOps device token
      description: >
        Identité propre à un appareil, remise une seule fois après confirmation
        Web et révocable sans fermer les autres appareils ni sessions Web.
    PairingBearer:
      type: http
      scheme: bearer
      bearerFormat: CairnOps pairing token
      description: >
        Code de 256 bits à usage unique et valable dix minutes. Il autorise
        seulement la revendication mobile et la lecture du résultat du même
        appairage.
    WebhookBearer:
      type: http
      scheme: bearer
      description: Secret de 256 bits affiché une seule fois lors de la création du webhook générique.
  schemas:
...
    DevicePairingInvitation:
      type: object
      required: [pairing, instance_url, token, qr_payload]
      properties:
        pairing: { $ref: '#/components/schemas/DevicePairing' }
        instance_url: { type: string, format: uri }
        token:
          type: string
          description: Secret de l'appairage affiché uniquement dans cette réponse.
        qr_payload:
          type: string
          description: URI cairnops://pair contenant l'URL manuelle et le secret temporaire.
...
paths:
...
  /api/v1/device-pairings:
    post:
      operationId: createDevicePairing
      description: >-
        Crée depuis le navigateur une invitation personnelle valable dix minutes.
        Le secret et l'URI du QR ne sont rendus qu'une fois et ne sont jamais
        persistés en clair.
      security: [{ SessionCookie: [] }]
      responses:
        '201':
          description: Invitation prête à être scannée
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DevicePairingInvitation' }
        '401': { description: Session Web requise }
        '403': { description: Origine Web non autorisée }
  /api/v1/device-pairings/claim:
    post:
      operationId: claimDevicePairing
      description: >-
        Revendique l'invitation depuis l'application mobile. Cette étape ne crée
        encore aucune identité : le navigateur doit montrer l'appareil puis le
        confirmer explicitement.
      security: [{ PairingBearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DevicePairingClaim' }
      responses:
        '202':
          description: Revendication en attente de confirmation Web
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DevicePairingResult' }
        '400': { description: Identité mobile invalide }
        '401': { description: Code d'appairage requis }
        '404': { description: Code inconnu }
        '409': { description: Code déjà revendiqué ou annulé }
        '410': { description: Code expiré }
  /api/v1/device-pairings/result:
    get:
      operationId: getDevicePairingResult
      description: >-
        Permet au mobile d'attendre la confirmation. Le jeton d'appareil est
        retiré de son stockage scellé dès cette réponse et ne peut jamais être
        demandé une seconde fois.
      security: [{ PairingBearer: [] }]
      responses:
        '200':
          description: État courant, éventuellement accompagné du jeton remis une fois
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DevicePairingResult' }
        '401': { description: Code d'appairage requis }
        '404': { description: Code inconnu }
        '409': { description: Appairage annulé }
        '410': { description: Code expiré ou jeton déjà retiré }
  /api/v1/device-pairings/{pairingID}:
    get:
      operationId: getDevicePairing
      description: Relit dans le navigateur l'étape et l'identité mobile revendiquée.
      security: [{ SessionCookie: [] }]
      parameters:
        - in: path
          name: pairingID
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Appairage appartenant à la personne connectée
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DevicePairing' }
        '400': { description: Identifiant invalide }
        '401': { description: Session Web requise }
        '404': { description: Appairage absent }
    delete:
      operationId: cancelDevicePairing
      description: Annule une invitation qui n'a pas encore été confirmée.
      security: [{ SessionCookie: [] }]
      parameters:
        - in: path
          name: pairingID
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: Invitation annulée }
        '400': { description: Identifiant invalide }
        '401': { description: Session Web requise }
        '403': { description: Origine Web non autorisée }
        '404': { description: Appairage absent ou déjà confirmé }
  /api/v1/device-pairings/{pairingID}/confirmation:
    post:
      operationId: confirmDevicePairing
      description: >-
        Confirme dans le navigateur l'appareil qui a revendiqué l'invitation et
        crée son identité révocable. Le jeton reste scellé jusqu'à son unique
        retrait par le mobile.
      security: [{ SessionCookie: [] }]
      parameters:
        - in: path
          name: pairingID
          required: true
          schema: { type: string, format: uuid }
      responses:
        '201':
          description: Identité d'appareil créée
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DevicePairing' }
        '400': { description: Identifiant invalide }
        '401': { description: Session Web requise }
        '403': { description: Origine Web non autorisée }
        '404': { description: Appairage absent }
        '409': { description: Scan absent, appairage annulé ou déjà confirmé }
        '410': { description: Appairage expiré }
  /api/v1/devices:
    get:
      operationId: listDevices
      description: >-
        Liste ses propres appareils ; un Administrateur voit ceux de tout
        l'Espace opérationnel. Aucun jeton, destinataire Push ni clé n'est rendu.
      security: [{ SessionCookie: [] }, { DeviceBearer: [] }]
      responses:
        '200':
          description: Appareils actifs et révoqués
          content:
            application/json:
              schema:
                type: object
                required: [devices]
                properties:
                  devices:
                    type: array
                    items: { $ref: '#/components/schemas/Device' }
        '401': { description: Authentification requise }
  /api/v1/devices/{deviceID}:
    patch:
      operationId: updateDevice
      description: >-
        Renomme un appareil, change sa préférence de contenu ou fait tourner sa
        clé et son destinataire opaque après un changement APNs ou FCM.
      security: [{ SessionCookie: [] }, { DeviceBearer: [] }]
      parameters:
        - in: path
          name: deviceID
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DeviceUpdate' }
      responses:
        '200':
          description: Appareil mis à jour
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Device' }
        '400': { description: Entrée invalide }
        '401': { description: Authentification requise }
        '403': { description: Origine Web non autorisée }
        '404': { description: Appareil absent }
        '409': { description: Appareil déjà révoqué }
    delete:
      operationId: revokeDevice
      description: >-
        Révoque immédiatement cette seule identité et annule ses livraisons Push
        encore en attente, sans toucher aux autres appareils ni sessions Web.
      security: [{ SessionCookie: [] }, { DeviceBearer: [] }]
      parameters:
        - in: path
          name: deviceID
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: Appareil révoqué }
        '400': { description: Identifiant invalide }
        '401': { description: Authentification requise }
        '403': { description: Origine Web non autorisée }
        '404': { description: Appareil absent ou déjà révoqué }
  /api/v1/events:
    get:
      operationId: connectEventStream
      description: >-
        Établit un WebSocket authentifié. Sans curseur, le premier message ready fixe la
        version courante. Avec after, les événements postérieurs sont rejoués dans l'ordre,
        puis diffusés au fil de leur enregistrement. Ce canal signale les changements ; les
        lectures et commandes restent dans l'API REST.
      security: [{ SessionCookie: [] }, { DeviceBearer: [] }]
      parameters:
        - in: query
          name: after
          schema: { type: integer, format: int64, minimum: 0 }
          description: Dernière version appliquée par le client avant une reconnexion.
      x-websocket-messages:
        $ref: '#/components/schemas/RealtimeMessage'
      responses:
        '101': { description: WebSocket établi }
        '400': { description: Curseur invalide }
        '401': { description: Authentification requise }
        '403': { description: Origine Web non autorisée }
        '503': { description: Suite d'événements indisponible }
  /api/v1/connectors:
    get:
      operationId: listConnectors
      security: [{ SessionCookie: [] }, { DeviceBearer: [] }]
      responses:
        '200':
          description: Connecteurs persistés sans leurs secrets
          content:
            application/json:
              schema:
                type: object
                required: [connectors]
                properties:
                  connectors:
                    type: array
                    items: { $ref: '#/components/schemas/Connector' }
        '401': { description: Authentification requise }
  /api/v1/connectors/{connectorID}/indicator-configuration:
    get:
      operationId: getIndicatorConfiguration
      description: Lit la configuration enregistrée, sans contacter le produit d'origine.
      security: [{ SessionCookie: [] }]
      parameters:
        - in: path
          name: connectorID
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Configuration et historique local, content: { application/json: { schema: { $ref: '#/components/schemas/IndicatorConfiguration' } } } }
        '401': { description: Authentification requise }
        '403': { description: Rôle Administrateur requis }
        '404': { description: Connecteur absent }
    put:
      operationId: applyIndicatorConfiguration
      description: >-
        Revalide l'aperçu distant puis applique atomiquement le périmètre, les
        sélections et les profils explicitement confirmés. Les points masqués
        ne sont pas effacés immédiatement : ils expirent selon la rétention.
      security: [{ SessionCookie: [] }]
      parameters:
        - in: path
          name: connectorID
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [bindings, profiles]
              properties:
                bindings: { type: array, maxItems: 5000, items: { $ref: '#/components/schemas/IndicatorBindingInput' } }
                profiles: { type: array, maxItems: 100, items: { $ref: '#/components/schemas/IndicatorProfileInput' } }
                summary: { type: string, maxLength: 500 }
      responses:
        '200': { description: Configuration appliquée, content: { application/json: { schema: { $ref: '#/components/schemas/IndicatorConfiguration' } } } }
        '400': { description: Sélection invalide ou à confirmer de nouveau }
        '401': { description: Authentification requise }
        '403': { description: Rôle Administrateur requis }
        '404': { description: Connecteur absent }
  /api/v1/connectors/{connectorID}/indicator-configuration/preview:
    post:
      operationId: previewIndicatorConfiguration
      description: Redécouvre les Cibles et le catalogue court permis par les capacités actuelles du produit d'origine.
      security: [{ SessionCookie: [] }]
      parameters:
        - in: path
          name: connectorID
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Aperçu temporaire à confirmer, content: { application/json: { schema: { $ref: '#/components/schemas/IndicatorConfiguration' } } } }
        '401': { description: Authentification requise }
        '403': { description: Rôle Administrateur requis }
        '404': { description: Connecteur absent ou suspendu }
  /api/v1/connectors/zabbix/preview:
    post:
      operationId: previewZabbixConnector
      description: Vérifie TLS, la version API et la portée du jeton, puis propose des rapprochements explicables par nom, IP ou nom DNS sans importer.
      security: [{ SessionCookie: [] }, { DeviceBearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [address, api_token]
              properties:
                name: { type: string, minLength: 1, maxLength: 160, default: Zabbix }
                address: { type: string, format: uri, maxLength: 2048 }
                api_token: { type: string, minLength: 1, maxLength: 4096, writeOnly: true }
      responses:
        '200':
          description: Aperçu temporaire sans modification des Cibles
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ZabbixPreview' }
        '400': { description: Entrée invalide }
        '401': { description: Authentification requise }
        '403': { description: Rôle Administrateur requis }
        '422': { description: Zabbix inaccessible, incompatible ou jeton insuffisant }
  /api/v1/connectors/zabbix/import:
    post:
      operationId: importZabbixConnector
      description: Revalide l'aperçu, scelle le jeton et importe uniquement les hôtes explicitement sélectionnés.
      security: [{ SessionCookie: [] }, { DeviceBearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [receipt, host_ids]
              properties:
                receipt: { type: string, writeOnly: true }
                host_ids:
                  type: array
                  minItems: 1
                  maxItems: 5000
                  uniqueItems: true
                  items: { type: string }
                target_assignments:
                  type: object
                  description: Affectations explicites des identifiants Zabbix sélectionnés vers des Cibles actives.
                  additionalProperties: { type: string, format: uuid }
      responses:
        '201':
          description: Connecteur enregistré et Cibles créées, réutilisées ou déjà liées
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ZabbixImport' }
        '400': { description: Sélection ou reçu invalide }
        '401': { description: Authentification requise }
        '403': { description: Rôle Administrateur requis }
        '410': { description: Aperçu expiré }
        '422': { description: La revalidation Zabbix a échoué }
  /api/v1/connectors/uptime-kuma/preview:
    post:
      operationId: previewUptimeKumaConnector
      description: Vérifie TLS, la clé API et la métrique monitor_status, puis propose des rapprochements explicables par nom, IP ou nom DNS sans importer.
      security: [{ SessionCookie: [] }, { DeviceBearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [address, api_key]
              properties:
                name: { type: string, minLength: 1, maxLength: 160, default: Uptime Kuma }
                address: { type: string, format: uri, maxLength: 2048 }
                api_key: { type: string, minLength: 1, maxLength: 4096, writeOnly: true }
      responses:
        '200':
          description: Aperçu temporaire des moniteurs sans modification des Cibles
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UptimeKumaPreview' }
        '400': { description: Entrée invalide }
        '401': { description: Authentification requise }
        '403': { description: Rôle Administrateur requis }
        '422': { description: Endpoint métriques inaccessible, invalide ou clé refusée }
  /api/v1/connectors/uptime-kuma/import:
    post:
      operationId: importUptimeKumaConnector
      description: Revalide l'aperçu, scelle la clé API et importe uniquement les moniteurs explicitement sélectionnés.
      security: [{ SessionCookie: [] }, { DeviceBearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [receipt, monitor_ids]
              properties:
                receipt: { type: string, writeOnly: true }
                monitor_ids:
                  type: array
                  minItems: 1
                  maxItems: 5000
                  uniqueItems: true
                  items: { type: string }
                target_assignments:
                  type: object
                  description: Affectations explicites des identifiants Kuma sélectionnés vers des Cibles actives.
                  additionalProperties: { type: string, format: uuid }
      responses:
        '201':
          description: Connecteur enregistré et Cibles créées, réutilisées ou déjà liées
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UptimeKumaImport' }
        '400': { description: Sélection ou reçu invalide }
        '401': { description: Authentification requise }
        '403': { description: Rôle Administrateur requis }
        '410': { description: Aperçu expiré }
        '422': { description: La revalidation Uptime Kuma a échoué }
...